调整IGUIButton下的按钮贴图 setUseAlphaChannel(true)

+++ Classes/gframe/deck_manager.h
@@ -44,7 +44,7 @@ public:
 	void LoadLFList(irr::android::InitOptions *options);
 	const wchar_t* GetLFListName(unsigned int lfhash);
 	const LFList* GetLFList(unsigned int lfhash);
-	unsigned int CheckDeck(const Deck& deck, unsigned int lfhash, int rule);
+	uint32_t CheckDeck(const Deck& deck, unsigned int lfhash, size_t rule);
 	bool LoadCurrentDeck(const wchar_t* file, bool is_packlist = false);
 	bool LoadCurrentDeck(irr::gui::IGUIComboBox* cbCategory, irr::gui::IGUIComboBox* cbDeck);
 	bool LoadCurrentDeck(std::istringstream& deckStream, bool is_packlist = false);
